misc: add additional type safety in `pandas` impl of `when`

+++ narwhals/_pandas_like/namespace.py
@@ -319,7 +319,9 @@ class PandasWhen:
             plx, self._otherwise_value, shape=condition.shape
         )._call(df)[0]
 
-        return [value_series.zip_with(condition, otherwise_series)]
+        return [
+            value_series.zip_with(condition, otherwise_series).cast(value_series.dtype)
+        ]
 
     def then(self, value: PandasLikeExpr | PandasLikeSeries | Any) -> PandasThen:
         self._then_value = value
